# improve speed of GFW calculations