--- a/R/calc_treecover_area.R
+++ b/R/calc_treecover_area.R
@@ -79,6 +79,9 @@ NULL
                                  verbose = TRUE,
                                  ...) {
   # initial argument checks
+  if (!requireNamespace("exactextractr", quietly = TRUE)){
+    stop("R package 'exactextractr' required. Please install via 'install.packages('exactextractr')'")
+  }
   # handling of return value if resources are missing, e.g. no overlap
   if (any(is.null(gfw_treecover), is.null(gfw_lossyear))) {
     return(NA)
@@ -88,12 +91,12 @@ NULL
 
   if (any(years < 2000)) {
     warning(paste("Cannot calculate treecover statistics ",
-      "for years smaller than 2000.",
-      sep = ""
+                  "for years smaller than 2000.",
+                  sep = ""
     ))
     years <- years[years >= 2000]
     if (length(years) == 0) {
-      return(tibble(years = NA, treecover = NA))
+      return(tibble(years = NA, treecover = NA, loss = NA))
     }
   }
 
@@ -101,14 +104,14 @@ NULL
   minmax_gfw_treecover <- unique(as.vector(minmax(gfw_treecover)))
   if (length(minmax_gfw_treecover) == 1) {
     if (minmax_gfw_treecover == 0 | is.nan(minmax_gfw_treecover)) {
-      return(tibble(years = years, treecover = rep(0, length(years))))
+      return(tibble(years = years, treecover = 0, loss = 0))
     }
   }
 
   # check additional arguments
   min_cover_msg <- paste("Argument 'min_cover' for indicator 'treecover' ",
-    "must be a numeric value between 0 and 100.",
-    sep = ""
+                         "must be a numeric value between 0 and 100.",
+                         sep = ""
   )
 
   if (is.numeric(min_cover)) {
@@ -121,103 +124,105 @@ NULL
   }
 
   min_size_msg <- paste("Argument 'min_size' for indicator 'treecover' must be ",
-    "anumeric value greater 0.",
-    sep = ""
+                        "anumeric value greater 0.",
+                        sep = ""
   )
   if (!is.numeric(min_size) | min_size <= 0) stop(min_size_msg, call. = FALSE)
 
-  # skip if the area of the
-
-  #----------------------------------------------------------------------------
-  # start calculation if everything is set up correctly
-  # retrieve an area raster
-  arearaster <- cellSize(
-    gfw_treecover,
-    unit = "ha"
-  )
-  # rasterize the polygon
-  polyraster <- rasterize(
-    vect(x), gfw_treecover,
-    field = 1, touches = TRUE
-  )
   # mask gfw_treecover
-  gfw_treecover <- mask(
-    gfw_treecover, polyraster
-  )
+  gfw_treecover <- mask(gfw_treecover, x)
+
+  # binarize the gfw_treecover layer based on min_cover argument
+  binary_gfw_treecover <- classify(gfw_treecover,
+                                   rcl = matrix(c(NA, NA, 0,
+                                                  0, min_cover, 0,
+                                                  min_cover, 100, 1), ncol = 3, byrow = TRUE),
+                                   include.lowest = TRUE)
+
+  # create patches
+  if(!requireNamespace("landscapemetrics", quietly = TRUE)){
+    message("Consider running `install.packages('landscapemetrics') to improve performance of GFW routines.")
+    patched <-  patches(binary_gfw_treecover, directions = 4, zeroAsNA = TRUE)
+  } else {
+    patched <- landscapemetrics::get_patches(binary_gfw_treecover, class = 1, direction = 4)[[1]][[1]]
+  }
 
   # mask lossyear
-  gfw_lossyear <- mask(
-    gfw_lossyear, polyraster
-  )
-  # binarize the gfw_treecover layer based on min_cover argument
-  binary_gfw_treecover <- classify(
-    gfw_treecover,
-    rcl = matrix(c(0, min_cover, 0, min_cover, 100, 1), ncol = 3, byrow = TRUE),
-    include.lowest = TRUE
-  )
-  # retrieve patches of comprehensive forest areas
-  patched <- patches(
-    binary_gfw_treecover,
-    directions = 4, zeroAsNA = TRUE
-  )
+  gfw_lossyear <- mask(gfw_lossyear, binary_gfw_treecover)
+  gfw_lossyear <- ifel(gfw_lossyear == 0, NA, gfw_lossyear)
 
-  unique_vals <- unique(as.vector(minmax(patched)))
-  if (length(unique_vals) == 1) {
-    if (is.nan(unique_vals)) {
-      return(tibble(years = years, treecover = rep(0, length(years))))
-    }
-  }
-  # get the sizes of the patches
-  patchsizes <- zonal(
-    arearaster, patched, sum,
-    as.raster = TRUE
-  )
-  # remove patches smaller than threshold
-  binary_gfw_treecover <- ifel(
-    patchsizes < min_size, 0, binary_gfw_treecover
-  )
-  # return 0 if binary gfw_treecover only consists of 0 or nan
-  minmax_gfw_treecover <- unique(as.vector(minmax(binary_gfw_treecover)))
-  if (length(minmax_gfw_treecover) == 1) {
-    if (minmax_gfw_treecover == 0 | is.nan(minmax_gfw_treecover)) {
-      return(
-        tibble(
+
+  gfw <- c(binary_gfw_treecover, gfw_lossyear, patched)
+  names(gfw) <- c("treecover", "lossyear", "patches")
+
+  gfw_stats <- exactextractr::exact_extract(
+    gfw, x, function(data, min_size){
+
+      # retain only forest pixels and set area to ha
+      data <- .prep_gfw_data(data, min_size)
+      losses <- .sum_gfw(data, "coverage_area")
+      names(losses)[2] <- "loss"
+      org_coverage <- sum(data[["coverage_area"]])
+
+      if(all(losses[["loss"]] == 0)) {
+        result <- data.frame(
           years = years,
-          treecover = rep(0, length(years))
+          treecover =  org_coverage,
+          loss = 0
         )
-      )
-    }
-  }
-  # set no loss occurrences to NA
-  gfw_lossyear <- ifel(
-    gfw_lossyear == 0, NA, gfw_lossyear
-  )
-  # exclude non-tree pixels from lossyear layer
-  gfw_lossyear <- mask(
-    gfw_lossyear, binary_gfw_treecover
-  )
+        return(result)
+      }
 
-  # get forest cover statistics for each year
-  yearly_cover_values <- lapply(years, function(y) {
-    y <- y - 2000
-    current_gfw_treecover <- ifel(
-      gfw_lossyear <= y, 0, binary_gfw_treecover
-    )
-    current_arearaster <- mask(
-      arearaster, current_gfw_treecover,
-      maskvalues = c(NA, 0)
-    )
-    ha_sum_gfw_treecover <- zonal(
-      current_arearaster, polyraster, sum,
-      na.rm = TRUE
-    )[2]
-    as.numeric(ha_sum_gfw_treecover)
-  })
+      previous_losses <- sum(losses[["loss"]][losses[["years"]] < years[1]])
 
-  # memory clean up
-  rm(arearaster, binary_gfw_treecover, patchsizes, patched, polyraster)
-  # return a data-frame
-  tibble(years = years, treecover = as.vector(unlist(yearly_cover_values)))
+      if (previous_losses != 0){
+        org_coverage <- org_coverage - previous_losses
+      }
+
+      losses <- losses[losses[["years"]] %in% years, ]
+      losses[["treecover"]] <- org_coverage
+      losses[["treecover"]] <- losses[["treecover"]] - cumsum(losses[["loss"]])
+      losses[ ,c("years", "treecover")]
+    }, min_size = min_size, coverage_area = TRUE, summarize_df = TRUE)
+
+  rm(gfw, binary_gfw_treecover, gfw_lossyear, patched)
+  gc()
+
+  tibble::as_tibble(gfw_stats)
+
+}
+
+
+.prep_gfw_data <- function(data, min_size) {
+
+  # retain only forest pixels and set area to ha
+  data <- data[data[["treecover"]] == 1, ]
+  data[["coverage_area"]] <- data[["coverage_area"]] / 10000 # to ha
+
+  # exclude patches smaller than threshold
+  patch_sizes <- by(data[["coverage_area"]], data[["patches"]], sum)
+  keep_patches <- names(patch_sizes)[which(patch_sizes > min_size)]
+  keep_patches <- as.numeric(keep_patches)
+
+  data <- data[data[["patches"]] %in% keep_patches, ]
+
+}
+
+
+.sum_gfw <- function(data, what = "coverage_area") {
+
+  # calculate loss area by year
+  df <- data.frame(years = 2000:2022, var = 0)
+  names(df)[2] <- what
+  my_sum <- by(data[[what]], data[["lossyear"]], sum, na.rm = TRUE)
+  sum_years <- as.numeric(names(my_sum))
+  sum_years <- sum_years + 2000
+
+  index <- match(sum_years, df[["years"]])
+  df[[what]][index] <- as.numeric(my_sum)
+
+  df
+
 }
 
 register_indicator(