--- a/tests/testthat/_snaps/calc_treecoverloss_emissions.md
+++ b/tests/testthat/_snaps/calc_treecoverloss_emissions.md
@@ -3,5 +3,5 @@
     Code
       result$emissions
     Output
-      [1]    0    7   46 1419  528  300
+      [1]    0    7   46 1419  528  295