… tests/testthat/_snaps/calc_treecover_area.md
… tests/testthat/_snaps/calc_treecover_area_and_emissions.md
… tests/testthat/test-calc_indicator.R