This PR reworks the GFW routines and improves the speed of calculation up to 10 times by introducing two main changes.
The first is that `exactectractr` is now required for the indicator calculations. It is still included in `SUGGESTS`, but users are informed to install it in case `requireNamespace()` returns `FALSE`.

The main speed-improvment, however,  is achieved by relying on [`landscapemetrics::get_patches()`](https://r-spatialecology.github.io/landscapemetrics/) instead of [`terra::patches()`](https://rspatial.github.io/terra/reference/patches.html). Thus, `landscapemetrics` is now also included in SUGGESTS, but the function code wil fallback to `terra` if it is not installed and issue a message to the user advising to install `landscapemetrics` for better computation times.