@goergen95 2023-12-19
Just to provide some evidence, here is a somewhat extrem use case. I simply took the bounding box of a large PA in Brazil (WDPAID = 33613) and compared the two routines. 

Results: The new routine is 51 times faster on my machine for this AOI and the difference in the treecover estimation is about 0.5% (3543 ha)! :tada:

Current routine:

``` r
remotes::install_github("mapme-initiative/mapme.biodiversity", ref = "main")
#> Skipping install of 'mapme.biodiversity' from a github remote, the SHA1 (7b0a2173) has not changed since last install.
#>   Use `force = TRUE` to force installation
library(sf)
#> Linking to GEOS 3.11.1, GDAL 3.8.1, PROJ 9.1.1; sf_use_s2() is TRUE
library(mapme.biodiversity)


aoi <- "POLYGON ((-62.2827 -13.5376, -61.1715 -13.5376, -61.1715 -12.8, -62.2827 -12.8, -62.2827 -13.5376))"
aoi <- st_as_sfc(aoi, crs = st_crs("EPSG:4326")) |> st_as_sf()
area <- st_area(aoi)
units::set_units(area, "km²")
#> 9867.781 [km²]

outdir <- tempfile()
dir.create(outdir)
aoi <- init_portfolio(aoi, years = 2000:2022, outdir = outdir)

aoi <- get_resources(aoi, resources = c("gfw_treecover", "gfw_lossyear"),
                     vers_treecover = "GFC-2022-v1.10",
                     vers_lossyear = "GFC-2022-v1.10")
#> Starting process to download resource 'gfw_treecover'........
#> Starting process to download resource 'gfw_lossyear'........

timing <- system.time(aoi <- calc_indicators(aoi, "treecover_area"))
#> Argument 'min_size' for resource 'treecover_area' was not specified. Setting to default value of '10'.
#> Argument 'min_cover' for resource 'treecover_area' was not specified. Setting to default value of '35'.
aoi$treecover_area[[1]]
#> # A tibble: 23 × 2
#>    years treecover
#>    <int>     <dbl>
#>  1  2000   610823.
#>  2  2001   610089.
#>  3  2002   608337.
#>  4  2003   600961.
#>  5  2004   597900.
#>  6  2005   592184.
#>  7  2006   590614.
#>  8  2007   589804.
#>  9  2008   588828.
#> 10  2009   588093.
#> # ℹ 13 more rows
timing
#>     user   system  elapsed 
#> 1065.052   17.670 1082.735
```

<sup>Created on 2023-12-19 with [reprex v2.0.2](https://reprex.tidyverse.org)</sup>


New routine:

``` r
remotes::install_github("mapme-initiative/mapme.biodiversity", ref = "speed-up-gfw-routines")
#> Skipping install of 'mapme.biodiversity' from a github remote, the SHA1 (2bb12db1) has not changed since last install.
#>   Use `force = TRUE` to force installation
library(sf)
#> Linking to GEOS 3.11.1, GDAL 3.8.1, PROJ 9.1.1; sf_use_s2() is TRUE
library(mapme.biodiversity)


aoi <- "POLYGON ((-62.2827 -13.5376, -61.1715 -13.5376, -61.1715 -12.8, -62.2827 -12.8, -62.2827 -13.5376))"
aoi <- st_as_sfc(aoi, crs = st_crs("EPSG:4326")) |> st_as_sf()
area <- st_area(aoi)
units::set_units(area, "km²")
#> 9867.781 [km²]

outdir <- tempfile()
dir.create(outdir)
aoi <- init_portfolio(aoi, years = 2000:2022, outdir = outdir)

aoi <- get_resources(aoi, resources = c("gfw_treecover", "gfw_lossyear"),
                     vers_treecover = "GFC-2022-v1.10",
                     vers_lossyear = "GFC-2022-v1.10")
#> Starting process to download resource 'gfw_treecover'........
#> Starting process to download resource 'gfw_lossyear'........

timing <- system.time(aoi <- calc_indicators(aoi, "treecover_area"))
#> Argument 'min_size' for resource 'treecover_area' was not specified. Setting to default value of '10'.
#> Argument 'min_cover' for resource 'treecover_area' was not specified. Setting to default value of '35'.
aoi$treecover_area[[1]]
#> # A tibble: 23 × 2
#>    years treecover
#>    <int>     <dbl>
#>  1  2000   614366.
#>  2  2001   613628.
#>  3  2002   611866.
#>  4  2003   604446.
#>  5  2004   601365.
#>  6  2005   595615.
#>  7  2006   594036.
#>  8  2007   593221.
#>  9  2008   592239.
#> 10  2009   591500.
#> # ℹ 13 more rows
timing
#>    user  system elapsed 
#>  16.653   4.430  21.083
```

<sup>Created on 2023-12-19 with [reprex v2.0.2](https://reprex.tidyverse.org)</sup>