@goergen95 2023-12-21
Here is another comparison when using a grid. Results indicate that processing time is reduced by half, differences in area estiamation for grid cell number 6 is about 1.7% (8ha).

Current routine:

``` r
remotes::install_github("mapme-initiative/mapme.biodiversity", ref = "main")
#> Skipping install of 'mapme.biodiversity' from a github remote, the SHA1 (7b0a2173) has not changed since last install.
#>   Use `force = TRUE` to force installation
library(sf)
#> Linking to GEOS 3.11.1, GDAL 3.8.1, PROJ 9.1.1; sf_use_s2() is TRUE
library(future)
library(progressr)
library(mapme.biodiversity)


aoi <- "POLYGON ((-62.2827 -13.5376, -61.1715 -13.5376, -61.1715 -12.8, -62.2827 -12.8, -62.2827 -13.5376))"
aoi <- st_as_sfc(aoi, crs = st_crs("EPSG:4326")) |> st_as_sf()
aoi <- st_make_grid(aoi, cellsize = c(0.025, 0.025)) |> st_as_sf()
area <- st_area(aoi)
mean(units::set_units(area[1], "km²"))
#> 7.513411 [km²]

outdir <- tempfile()
dir.create(outdir)
aoi <- init_portfolio(aoi, years = 2000:2022, outdir = outdir)

aoi <- get_resources(aoi, resources = c("gfw_treecover", "gfw_lossyear"),
                     vers_treecover = "GFC-2022-v1.10",
                     vers_lossyear = "GFC-2022-v1.10")
#> Starting process to download resource 'gfw_treecover'........
#> Starting process to download resource 'gfw_lossyear'........


plan(multisession, workers = 6)
with_progress({
  timing <- system.time(aoi <- calc_indicators(aoi, "treecover_area"))
})
#> Argument 'min_size' for resource 'treecover_area' was not specified. Setting to default value of '10'.
#> Argument 'min_cover' for resource 'treecover_area' was not specified. Setting to default value of '35'.
plan(sequential)
aoi$treecover_area[[6]]
#> # A tibble: 23 × 2
#>    years treecover
#>    <int>     <dbl>
#>  1  2000      452.
#>  2  2001      452.
#>  3  2002      450.
#>  4  2003      450.
#>  5  2004      450.
#>  6  2005      450.
#>  7  2006      450.
#>  8  2007      450.
#>  9  2008      450.
#> 10  2009      450.
#> # ℹ 13 more rows
timing
#>    user  system elapsed 
#>   6.804   0.312 172.218
```

<sup>Created on 2023-12-21 with [reprex v2.0.2](https://reprex.tidyverse.org)</sup>


New routine:

``` r
remotes::install_github("mapme-initiative/mapme.biodiversity", ref = "speed-up-gfw-routines")
#> Skipping install of 'mapme.biodiversity' from a github remote, the SHA1 (2bb12db1) has not changed since last install.
#>   Use `force = TRUE` to force installation
library(sf)
#> Linking to GEOS 3.11.1, GDAL 3.8.1, PROJ 9.1.1; sf_use_s2() is TRUE
library(future)
library(progressr)
library(mapme.biodiversity)


aoi <- "POLYGON ((-62.2827 -13.5376, -61.1715 -13.5376, -61.1715 -12.8, -62.2827 -12.8, -62.2827 -13.5376))"
aoi <- st_as_sfc(aoi, crs = st_crs("EPSG:4326")) |> st_as_sf()
aoi <- st_make_grid(aoi, cellsize = c(0.025, 0.025)) |> st_as_sf()
area <- st_area(aoi)
mean(units::set_units(area[1], "km²"))
#> 7.513411 [km²]

outdir <- tempfile()
dir.create(outdir)
aoi <- init_portfolio(aoi, years = 2000:2022, outdir = outdir)

aoi <- get_resources(aoi, resources = c("gfw_treecover", "gfw_lossyear"),
                     vers_treecover = "GFC-2022-v1.10",
                     vers_lossyear = "GFC-2022-v1.10")
#> Starting process to download resource 'gfw_treecover'........
#> Starting process to download resource 'gfw_lossyear'........


plan(multisession, workers = 6)
with_progress({
  timing <- system.time(aoi <- calc_indicators(aoi, "treecover_area"))
})
#> Argument 'min_size' for resource 'treecover_area' was not specified. Setting to default value of '10'.
#> Argument 'min_cover' for resource 'treecover_area' was not specified. Setting to default value of '35'.
plan(sequential)
aoi$treecover_area[[6]]
#> # A tibble: 23 × 2
#>    years treecover
#>    <int>     <dbl>
#>  1  2000      444.
#>  2  2001      444.
#>  3  2002      443.
#>  4  2003      443.
#>  5  2004      443.
#>  6  2005      443.
#>  7  2006      443.
#>  8  2007      443.
#>  9  2008      443.
#> 10  2009      443.
#> # ℹ 13 more rows
timing
#>    user  system elapsed 
#>   3.717   0.138  88.381
```

<sup>Created on 2023-12-21 with [reprex v2.0.2](https://reprex.tidyverse.org)</sup>